programming/yaml-json-guide/yaml-json-guide_files/requirements.txt
# YAML processing
PyYAML>=6.0
ruamel.yaml>=0.17.0

# JSON processing
jsonschema>=4.17.0

# Validation
yamllint>=1.32.0

# CLI tools
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
